Coding 005

+++ Assets/Systems/GameController/GameController.cs
@@ -1,6 +1,9 @@
 using System.Collections;
 using System.Collections.Generic;
+using DG.Tweening;
+using TMPro;
 using UnityEngine;
+using UnityEngine.SceneManagement;
 
 public class GameController : MonoBehaviour
 {
@@ -14,11 +17,37 @@ public class GameController : MonoBehaviour
     public float preFlipTime;
     public bool stackMode;
     private bool playing;
+    public float quota;
+    public float round;
+    public TextMeshProUGUI quotaText;
+    public GameObject roundWon;
+    private bool roundWonShowing;
+    public TextMeshProUGUI inGameQuotaText;
+    public TextMeshProUGUI inGameRoundCounter;
+    private int roundTillDeath;
+    private float totalPoints;
+    public TextMeshProUGUI currentPointCounter;
+
+    void Start()
+    {
+        roundTillDeath = 5;
+        currentPointCounter.text = "Points: 0";
+        round = 0;
+        quota = 15;
+        roundWonShowing = false;
+        startGame();
+    }
 
     public void startGame()
     {
         if (!playing)
         {
+            inGameRoundCounter.text = "Round: " + roundTillDeath.ToString();
+            inGameQuotaText.text = "Quota: " + quota.ToString();
+            if (roundWonShowing)
+            {
+                roundWon.GetComponent<CanvasGroup>().DOFade(0f, 0.5f).OnComplete(() => { roundWon.SetActive(false); roundWonShowing = false; });
+            }
             playing = true;
             StartCoroutine(setupCards());
         }
@@ -44,17 +73,29 @@ public class GameController : MonoBehaviour
 
     public void countPoints()
     {
-        points = 0;
+        for (int i = 0; i < hand.Count; i++)
+        {
+            if (hand[i].currentCard == null)
+            {
+                return;
+            }
+        }
         for (int i = 0; i < hand.Count; i++)
         {
             CardPlacement card = hand[i];
-            int neighboringCards = checkNextInList(i, hand[i].currentCard.GetComponent<CardObject>().points);
-            points += (1 + neighboringCards) * neighboringCards;
-            i += neighboringCards - 1;
+            if (card.currentCard != null)
+            {
+                int neighboringCards = checkNextInList(i, hand[i].currentCard.GetComponent<CardObject>().points);
+                points += (1 + (neighboringCards - 1)) * neighboringCards;
+                i += neighboringCards - 1;
+            }
         }
         foreach (CardPlacement handI in hand)
         {
-            handI.currentCard.GetComponent<CardObject>().delete();
+            if (handI.currentCard != null)
+            {
+                handI.currentCard.GetComponent<CardObject>().delete();
+            }
         }
         if (!stackMode)
         {
@@ -66,7 +107,39 @@ public class GameController : MonoBehaviour
                 }
             } // take out for stack mode
         }
+        totalPoints += points;
+        currentPointCounter.text = "Points: " + points.ToString();
         playing = false;
+        if (roundTillDeath == 1)
+        {
+            if (points >= quota)
+            {
+                roundTillDeath = 5;
+                roundWonShowing = true;
+                round++;
+                quotaText.text = points + "/" + quota;
+                quota = Mathf.Ceil(quota + (points / 2));
+                points = 0;
+                currentPointCounter.text = "Points: " + points.ToString();
+                roundWon.GetComponent<CanvasGroup>().alpha = 0;
+                roundWon.SetActive(true);
+                roundWon.GetComponent<CanvasGroup>().DOFade(1, 0.5f);
+            }
+            else
+            {
+                SceneManager.LoadScene("MenuScene");
+            }
+        }
+        else
+        {
+            roundTillDeath--;
+            roundWonShowing = true;
+            round++;
+            quotaText.text = points + "/" + quota;
+            roundWon.GetComponent<CanvasGroup>().alpha = 0;
+            roundWon.SetActive(true);
+            roundWon.GetComponent<CanvasGroup>().DOFade(1, 0.5f);
+        }
     }
     public void calculateCurrentSize()
     {
